+++ src/Server/Index.js
@@ -3,6 +3,9 @@ import path from 'path';
 import fs from 'fs';
 import showdown from 'showdown';
 
+import Utils from '../Utils/Utils';
+import Proxy from '../Server/Proxy';
+
 const converter = new showdown.Converter();
 
 async function setIndexMiddleware(expressApp, config) {
@@ -10,23 +13,9 @@ async function setIndexMiddleware(expressApp, config) {
   setLaunchpadMiddleware(expressApp, config);
 }
 
-function getOptionsVersion(ui5toolsData = {}, { ui5Version }) {
-  let ui5VersionArray = ui5Version.split('.');
-  if (ui5VersionArray.length >= 2) {
-    if (ui5VersionArray[1] >= 42) {
-      ui5toolsData.showTree = true;
-    }
-
-    if (ui5VersionArray[1] >= 65) {
-      ui5toolsData.theme = 'sap_fiori_3';
-    } else if (ui5VersionArray[1] >= 44) {
-      ui5toolsData.theme = 'sap_belize';
-    }
-  }
-  return ui5toolsData;
-}
+// SERVER INDEX
 function setServerIndexMiddleware(expressApp, config) {
-  let { ui5ToolsPath, folders, serverName, baseDir, manifests, ui5Version } = config;
+  let { ui5ToolsPath, folders, serverName, baseDir, manifests } = config;
   let ui5toolsData = {
     readme: '',
     folders: folders,
@@ -34,24 +23,30 @@ function setServerIndexMiddleware(expressApp, config) {
     serverName: serverName,
     launchpad: isLaunchpadMounted(config),
     docs: [],
-    showTree: false,
-    theme: 'sap_bluecrystal',
   };
-  getOptionsVersion(ui5toolsData, config);
-  // SERVER INDEX
+  Utils.getOptionsVersion(ui5toolsData, config);
+
+  let indexPath = path.join(ui5ToolsPath, 'index', 'ui5', 'webapp');
   let indexHTML = function (req, res, next) {
-    res.render(path.join(ui5ToolsPath, 'index', 'ui5', 'webapp', 'index'), { theme: ui5toolsData.theme });
+    res.render(path.join(indexPath, 'index'), { theme: ui5toolsData.theme });
   };
+  // render index with correct theme
   expressApp.get('/', indexHTML);
   expressApp.get('/index.html', indexHTML);
+
+  // render view with correct list or tree
   expressApp.get('/view/main.view.xml', function (req, res, next) {
     res.setHeader('content-type', 'text/xml');
-    res.render(path.join(ui5ToolsPath, 'index', 'ui5', 'webapp', 'view', 'main'), {
+    res.render(path.join(indexPath, 'view', 'main'), {
       showTree: ui5toolsData.showTree,
       launchpad: ui5toolsData.launchpad,
     });
   });
+
+  // Serve app files
   expressApp.use('/', express.static(path.join(ui5ToolsPath, 'index', 'ui5', 'webapp')));
+
+  // Serve app data
   expressApp.get('/ui5tools.json', function (req, res) {
     if (fs.existsSync(path.join(baseDir, 'README.md'))) {
       let readmeMD = fs.readFileSync(path.join(baseDir, 'README.md'), 'utf8');
@@ -102,13 +97,32 @@ function isLaunchpadMounted({ resourcesProxy }) {
 function setLaunchpadMiddleware(expressApp, config) {
   if (isLaunchpadMounted(config)) {
     // LAUNCHPAD IN /flp/
-    let { ui5ToolsPath, baseDir, manifests } = config;
+    let { ui5ToolsPath, baseDir, manifests, ui5Version } = config;
+
+    // DONT MOUNT RESOURCE ROOTS TO SIMULATE LAUNCHPAD
+    // Object.entries(manifests).forEach(([folder, manifest]) => {
+    //   fioriSandboxConfig.modulePaths[manifest['sap.app'].id] = `../${folder}`;
+    // });
+
+    let ui5toolsData = Utils.getOptionsVersion();
+    let flpPath = path.join(ui5ToolsPath, 'index', 'flp');
+
+    let indexFLP = function (req, res, next) {
+      res.render(path.join(flpPath, 'index'), { theme: ui5toolsData.theme });
+    };
+    expressApp.get('/flp/', indexFLP);
+    expressApp.get('/flp/index.html', indexFLP);
+
+    // expressApp.use('/flp/test-resources/sap/ushell/bootstrap/sandbox.js', function (req, res) {
+    //   res.sendFile(path.join(flpPath, 'sandbox.js'));
+    // });
+
     let fioriSandboxConfig = {
       modulePaths: {},
       applications: {},
     };
     Object.entries(manifests).forEach(([folder, manifest]) => {
-      let hash = folder.split('_').join('-').toLowerCase();
+      let hash = 'ui5tools-' + folder.toLowerCase();
       fioriSandboxConfig.applications[hash] = {
         additionalInformation: `SAPUI5.Component=${manifest['sap.app'].id}`,
         applicationType: 'SAPUI5',
@@ -117,26 +131,15 @@ function setLaunchpadMiddleware(expressApp, config) {
         title: folder,
       };
     });
-    // DONT MOUNT RESOURCE ROOTS TO SIMULATE LAUNCHPAD
-    // Object.entries(manifests).forEach(([folder, manifest]) => {
-    //   fioriSandboxConfig.modulePaths[manifest['sap.app'].id] = `../${folder}`;
-    // });
-    let ui5toolsData = getOptionsVersion({}, config);
-    let flpPath = path.join(ui5ToolsPath, 'index', 'flp', 'webapp');
-    let indexFLP = function (req, res, next) {
-      res.render(path.join(flpPath, 'index'), { theme: ui5toolsData.theme });
-    };
-    expressApp.get('/flp/', indexFLP);
-    expressApp.get('/flp/index.html', indexFLP);
-    expressApp.use('/flp/test-resources/sap/ushell/bootstrap/sandbox.js', function (req, res) {
-      res.sendFile(path.join(flpPath, 'sandbox.js'));
-    });
+
     expressApp.get('/flp/test-resources/sap/ushell/shells/sandbox/fioriSandboxConfig.json', function (req, res) {
       res.json(fioriSandboxConfig);
     });
     expressApp.get('/appconfig/fioriSandboxConfig.json', function (req, res) {
       res.sendFile(path.join(baseDir, 'fioriSandboxConfig.json'));
     });
+
+    Proxy.setTestResourcesProxy(expressApp, config);
   }
 }